Rename reshard_program_factory files to reshard_common as requested by Pavlo
- Rename reshard_program_factory.cpp to reshard_common.cpp
- Rename reshard_program_factory.hpp to reshard_common.hpp
- Update all includes to use the new header name
- Update CMakeLists.txt to reference the new filename

These files contain substantial utility functions and data structures that should be renamed rather than removed:
- Data structures: PageRange, Stride, PageStride, CorePageRange, CorePageStride
- Utility functions: get_core_page_ranges, get_runtime_args_for_given_ranges, reshard_multi_core
