+++ ttnn/cpp/ttnn/operations/data_movement/CMakeLists.txt
@@ -94,7 +94,7 @@ target_sources(
         sharded/interleaved_to_sharded/device/interleaved_to_sharded_program_factory.cpp
         sharded/interleaved_to_sharded/interleaved_to_sharded.cpp
         sharded/reshard/device/reshard_op.cpp
-        sharded/reshard/device/reshard_program_factory.cpp
+        sharded/reshard/device/reshard_common.cpp
         sharded/reshard/device/reshard_same_width_program_factory.cpp
         sharded/reshard/device/reshard_generic_program_factory.cpp
         sharded/reshard/device/reshard_same_height_program_factory.cpp
